# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Wrenn API
  description: AI agent execution platform API.
  version: "0.2.0"
servers:
  - url: http://localhost:8080
    description: Local development
security: []
paths:
...
  /v1/admin/users/{id}/admin:
    put:
      summary: Grant or revoke platform admin
      operationId: setUserAdmin
      tags: [admin]
      description: |
        Sets the platform admin flag on a user. Cannot remove the last admin.
        Requires platform admin access. Session caches for the target user
        are invalidated immediately so the flag flip takes effect on the
        user's next request.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "usr-a1b2c3d4"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [admin]
              properties:
                admin:
                  type: boolean
                  description: true to grant admin, false to revoke.
      responses:
        "204":
          description: Admin status updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Caller is not a platform admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/admin/users:
    get:
      summary: List all users (admin)
      operationId: adminListUsers
      tags: [admin]
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items: {type: object}
...
components:
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for capsule lifecycle operations. Create via POST /v1/api-keys.

    sessionAuth:
      type: apiKey
      in: cookie
      name: wrenn_sid
      description: |
        Opaque session cookie set by POST /v1/auth/login, /v1/auth/activate, or
        the OAuth callback. HttpOnly, Secure, SameSite=Strict. Idle window 6h,
        absolute lifetime 24h. State-changing requests also require an
        X-CSRF-Token header matching the wrenn_csrf cookie (double-submit).
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Double-submit CSRF token whose value must match the wrenn_csrf cookie.
        Required on all non-GET requests authenticated via session cookie.
        Not required for API key auth.

    hostTokenAuth:
      type: apiKey
      in: header
      name: X-Host-Token
      description: Host JWT returned from POST /v1/hosts/register or POST /v1/hosts/auth/refresh. Valid for 7 days.
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...